Converted to Maven project
