+++ src/main/java/com/github/idris/raytracer/pigments/Pigment.java
@@ -1,0 +1,10 @@
+package com.github.idris.raytracer.pigments;
+
+import com.github.idris.raytracer.Point;
+
+import java.awt.Color;
+
+
+public interface Pigment {
+	public Color getColor(Point p);
+}
