--- src/raytracer/pigments/Pigment.java
@@ -1,10 +1,0 @@
-package raytracer.pigments;
-
-import raytracer.Point;
-
-import java.awt.Color;
-
-
-public interface Pigment {
-	public Color getColor(Point p);
-}
